+++ app/(home)/page.tsx
@@ -2,8 +2,7 @@ import { FaqsSection } from "@/(features)/hero/faqs";
 import { Hero } from "@/(features)/hero";
 import { Opportunity } from "@/(features)/hero/opportunity";
 import { Sponsors } from "@/(features)/hero/sponsors";
-import AboutStats from "./_components/visuals/AboutStats";
-import About from "./_components/visuals/About";
+import { AboutStats, About } from "@/(features)/hero";
 
 export default function Home() {
   return (
