+++ src/(features)/gallery/data/gallery-photos/galleryPhotos2023.ts
