+++ src/(features)/gallery/index.ts
@@ -1,0 +1,1 @@
+export { default as Gallery } from "@/(features)/gallery/components/gallery";
