+++ src/(features)/hero/components/visuals/About.tsx
