+++ src/(features)/hero/components/visuals/AboutStats.tsx
