+++ src/(features)/hero/components/visuals/Skyline2.tsx
