+++ src/(features)/hero/index.ts
@@ -1,1 +1,3 @@
 export { default as Hero } from "@/(features)/hero/components/Hero";
+export { default as AboutStats } from "@/(features)/hero/components/visuals/AboutStats";
+export { default as About } from "@/(features)/hero/components/visuals/About";
