+++ src/(features)/winners/components/SpecialAwards.tsx
@@ -1,4 +1,4 @@
-import { Winner } from "@/types/WinnerType";
+import { Winner } from "@/(features)/winners";
 import Link from "next/link";
 import { slugify } from "@/utils/slugify";
