+++ src/(features)/winners/index.ts
@@ -1,0 +1,8 @@
+export { default as WinnersPodium } from "@/(features)/winners/components/WinnersPodium";
+export { default as YearDropdown } from "@/(features)/winners/components/YearDropdown";
+export { default as SpecialAwards } from "@/(features)/winners/components/SpecialAwards";
+export { pastWinners } from "@/(features)/winners/data/pastWinners";
+export type {
+  Winner,
+  PastWinners,
+} from "@/(features)/winners/types/WinnerType";
