+++ src/(features)/winners/types/WinnerType.ts
